requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.0.2
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
requests==2.23.0
urllib3==1.25.9
Werkzeug==1.0.1
